requirements.txt
asgiref==3.8.1
Django==5.1.2
django-appconf==1.0.6
django-ckeditor-5==0.2.15
django-cors-headers==4.5.0
django-filter==24.3
django-modeltranslation==0.19.11
django-parler==2.3
django-parler-rest==2.2
django-select2==8.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
